refactor: 가독성을 높이기 위해 코드를 별도의 클래스로 분리
